refactored css to comply to child components + cleaned up index.tsx
