+++ styles/Home.module.css
@@ -8,72 +8,11 @@
   background-color: white;
 }
 
-.sidebar {
-  background-color: lightblue;
-  width: 10%;
-  min-width: 120px;
-}
-
-.sbLogo {
-  width: 100%;
-  padding: 2%;
-  height: auto;
-  border-radius: 1rem;
-}
-
-.sbContent {
-  height: 70%;
-  margin: 10%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.splashButton {
-  background-color: gray;
-  border-color: rgba(128, 128, 128, 0.748);
-  margin: 5%;
-  border-radius: .3rem;
-}
-
-.secondButton {
-  background-color: gray;
-  border-color: rgba(128, 128, 128, 0.748);
-  margin: 5%;
-}
-
-.thirdButton {
-  background-color: gray;
-  border-color: rgba(128, 128, 128, 0.748);
-  margin: 5%;
-}
-
-.sbLinks {
-  margin: 10%;
-  display: block;
-  font-size: 100%;
-
-}
-
 .networkContainer {
   /* background-color: lightpink; */
   width: 90%;
   padding: 1rem;
   border: solid 1px grey;
-}
-
-.mainWaterfall {
-  /* background-color: lightseagreen; */
-  border: solid 1px grey;
-  height: 15%;
-}
-
-.detailList {
-  background-color: grey;
-  border: solid 1px grey;
-  border-top: none;
-  height: 85%;
-  /* overflow-y: scroll; */
 }
 
 .MuiBox-root {
