+++ styles/MainWaterfall.module.css
@@ -1,0 +1,5 @@
+.mainWaterfall {
+    /* background-color: lightseagreen; */
+    border: solid 1px grey;
+    height: 15%;
+  }
